Fixed synchronization issue with multiple vtkSynchronizedRenderers instances.
When multiple process groups were connected using multiple
vtkSynchronizedRenderers e.g. one for client-server and other for server-root
and server-satellites, we had a race condition. It was critical that the
server-slave-start-render happened before server-root-master-start-render. In
other words, it was necessary that the server-root first obtained the renderer
parameters from the client before relaying them to the satellites. Fixed that.